+++ server/v2/cometbft/abci.go
@@ -47,7 +47,7 @@ var _ abci.Application = (*Consensus[transaction.Tx])(nil)
 type Consensus[T transaction.Tx] struct {
 	logger           log.Logger
 	appName, version string
-	app              *appmanager.AppManager[T]
+	app              appmanager.AppManager[T]
 	appCloser        func() error
 	txCodec          transaction.Codec[T]
 	store            types.Store
@@ -82,7 +82,7 @@ type Consensus[T transaction.Tx] struct {
 func NewConsensus[T transaction.Tx](
 	logger log.Logger,
 	appName string,
-	app *appmanager.AppManager[T],
+	app appmanager.AppManager[T],
 	appCloser func() error,
 	mp mempool.Mempool[T],
 	indexedEvents map[string]struct{},
@@ -247,7 +247,7 @@ func (c *Consensus[T]) Query(ctx context.Context, req *abciproto.QueryRequest) (
 }
 
 func (c *Consensus[T]) maybeRunGRPCQuery(ctx context.Context, req *abci.QueryRequest) (resp *abciproto.QueryResponse, isGRPC bool, err error) {
-	// if this fails  then we cannot serve queries anymore
+	// if this fails then we cannot serve queries anymore
 	registry, err := c.getProtoRegistry()
 	if err != nil {
 		return nil, false, err