+++ store/v2/store.go
@@ -70,7 +70,7 @@ type RootStore interface {
 // Backend defines the interface for the RootStore backends.
 type Backend interface {
 	// GetStateStorage returns the SS backend.
-	GetStateStorage() VersionedDatabase
+	GetStateStorage() VersionedWriter
 
 	// GetStateCommitment returns the SC backend.
 	GetStateCommitment() Committer